fix: Address code review feedback for SSR/RSC implementation
This commit addresses priority fixes from external code review:

- Fix singleResult bug in hydrated data path - now properly returns single object vs array
- Remove unused staleTime option from prefetchLiveQuery
- Split server.tsx into server.ts (server-only) and hydration.tsx ('use client') for proper RSC module boundaries
- Add test coverage for singleResult with hydrated data
- Improve option detection from fragile duck typing to explicit property check
- Add comprehensive serialization constraints documentation to README

All 68 tests passing.
